+++ rephi_auto_chart/__init__.py
@@ -1,3 +1,3 @@
 """Re:PhiEdit Auto Chart Assistant."""
 
-__version__ = "2.5.1"
+__version__ = "2.5.2"